Refactor: python, cts: Move StopOnebyOne into the _cts module.
